+++ src/App.tsx
@@ -2,10 +2,9 @@ import React from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
 import Home from "./pages/Home";
 import Login from "./pages/Login";
-import Receive from "./pages/Receive/Receive";
 import Register from "./pages/Register";
 import Send from "./pages/Send/Send";
-import { useIntervalToGetLocation } from "./lib/hooks/locationHooks";
+import { useIntervalToGetLocation } from "./utils/location";
 
 function InValidateURL() {
   return <Navigate to="/" />;
@@ -18,7 +17,6 @@ function App() {
     <Routes>
       <Route path="/" index element={<Home />} />
       <Route path="/send" element={<Send />} />
-      <Route path="/receive" element={<Receive />} />
       <Route path="/login" element={<Login />} />
       <Route path="/register" element={<Register />} />
       <Route path="*" element={<InValidateURL />} />
