+++ src/utils/location.ts
@@ -1,5 +1,5 @@
 import { useEffect, useLayoutEffect, useRef, useState } from "react";
-import { TLLCoordinates } from "../lib/types/locationTypes";
+import { TLLCoordinates } from "../types/locationTypes";
 import { useMyPositionStore } from "../store/useMyPositionStore";
 
 export const useGetCurrentLocation = (options: PositionOptions = {}) => {
